Add measurement voting ambiguity resolution
This draft PR introduces an ambiguity resolution in the track finding
algorithm that takes two arguments, namely the number of votes per
measurement N and the required voting fraction per track $q$. It works
simply; first, we compute the $N$ best (highest p-value) tracks per
measurement, and then each measurement votes on the up-to $N$ tracks
that are most likely to belong to that measurement. A track is then kept
if out its $K$ track states, it receives at least $qK$ votes. In other
words, if for fraction $q$ of the track's states it is one of the $N$
most likely tracks, we keep it.